Rename TaskContainer.get(Class, String) to TaskCollection.named(String)
